requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
faiss-cpu==1.7.4
numpy==1.26.0
python-dotenv==1.0.0
openai==0.27.8

